requirements.txt
streamlit==1.28.1
pymongo==4.6.0
ollama==0.1.7
openai-whisper==20231117
speechrecognition==3.10.0
moviepy==1.0.3
pydub==0.25.1
apscheduler==3.10.4
pandas==2.1.3
fpdf2==2.7.6
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
